Downgrade JRuby from 9.1.2.0 to 9.0.5.0

+++ jruby/lib/ruby/stdlib/jars/classpath.rb
@@ -6,21 +6,6 @@ module Jars
 
   class Classpath
 
-    # convenient method
-    def self.require( scope = nil )
-      new.require( scope )
-    end
-
-    # convenient method
-    def self.classpath( scope = nil )
-      new.classpath( scope )
-    end
-
-    # convenient method
-    def self.classpath_string( scope = nil )
-      new.classpath_string( scope )
-    end
-
     def initialize( spec = nil, deps = nil )
       @spec = spec
       @deps = deps
@@ -36,11 +21,7 @@ module Jars
     end
 
     def dependencies_list
-      if @deps.nil?
-        deps = Jars.lock_path( mvn.basedir )
-        @deps = deps if deps && File.exist?( deps )
-      end
-      if @deps
+      if @deps ||= Jars.lock_path( mvn.basedir )
         @deps
       else
         resolve_dependencies
@@ -51,7 +32,7 @@ module Jars
     DEPENDENCY_LIST = 'dependencies.list'
     def resolve_dependencies
       basedir = workdir( 'pkg' ) || workdir( 'target' ) || workdir( '' )
-      deps = File.join( basedir, DEPENDENCY_LIST )
+      deps = File.join( basedir, DEPENDENCY_LIST )      
       mvn.resolve_dependencies_list( deps )
       deps
     end
@@ -84,8 +65,7 @@ module Jars
       deps = dependencies_list
       Lock.new( deps ).process( scope, &block )
     ensure
-      # just delete the temporary file if it exists
-      FileUtils.rm_f( DEPENDENCY_LIST )
+      FileUtils.rm_f( DEPENDENCY_LIST ) if deps
     end
     private :process
